+++ src/types/PipelineJson/PromptDialogJson.ts
@@ -12,5 +12,5 @@ export type PromptDialogJson = PromptTemplateJsonCommon & {
 
 /**
  * TODO: [🧠][🥜]
- * TODO: [🍙] Make some standart order of json properties
+ * TODO: [🍙] Make some standard order of json properties
  */