+++ src/types/PipelineJson/SimpleTemplateJson.ts
@@ -11,5 +11,5 @@ export type SimpleTemplateJson = PromptTemplateJsonCommon & {
 };
 
 /**
- * TODO: [🍙] Make some standart order of json properties
+ * TODO: [🍙] Make some standard order of json properties
  */