+++ src/utils/serialization/clonePipeline.ts
@@ -41,5 +41,5 @@ export function clonePipeline(pipeline: PipelineJson): PipelineJson {
 }
 
 /**
- * TODO: [🍙] Make some standart order of json properties
+ * TODO: [🍙] Make some standard order of json properties
  */